saving program

+++ StatesProgram/include/drive.h
@@ -1,2 +1,2 @@
-void runDrivetrain(vex::directionType rightDirection, vex::directionType leftDirection, float rightPower, float leftPower);
+void runDrivetrain(vex::directionType rightDirection, vex::directionType leftDirection, double rightPower, double leftPower);
 void holdDrivetrain();
